% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{available_models}
\alias{available_models}
\title{Available models and data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 18 rows and 9 columns.
}
\usage{
available_models
}
\description{
Models and data available in the app
}
\keyword{datasets}
