% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pathology.R
\name{filter_pathology}
\alias{filter_pathology}
\title{Filter pathology data}
\usage{
filter_pathology(data, phenotype, mouse_line, tissue)
}
\arguments{
\item{data}{Input data (e.g. \code{\link{phenotypes}})}

\item{phenotype}{Phenotype to filter by. Keeps data where \code{.data[["phenotype"]] \%in\% phenotype}}

\item{mouse_line}{Mouse lines to filter by. Keeps data where \code{.data[["mouse_line"]] \%in\% mouse_line}}

\item{tissue}{Tissue to filter by. Keeps data where \code{.data[["tissue"]] \%in\% tissue}}
}
\description{
Filter pathology data to only include data from the specified \code{phenotype}, \code{mouse_line}, and \code{tissue} values.
}
\examples{
phenotypes \%>\%
  filter_pathology(phenotype = "Plaque #", mouse_line = "BL6", tissue = "hippocampus")
}
