% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_expression_genes}
\alias{gene_expression_genes}
\title{Gene expressions genes}
\format{
An object of class \code{character} of length 55773.
}
\usage{
gene_expression_genes
}
\description{
A data set containing the genes in the gene expressions data.
}
\keyword{datasets}
