% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_expression_mouse_lines}
\alias{gene_expression_mouse_lines}
\title{A data set containing the mouse lines in the gene expressions data.}
\format{
An object of class \code{character} of length 3.
}
\usage{
gene_expression_mouse_lines
}
\description{
A data set containing the mouse lines in the gene expressions data.
}
\keyword{datasets}
