% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_expression_tissues}
\alias{gene_expression_tissues}
\title{A data set containing the tissues in the gene expressions data.}
\format{
An object of class \code{character} of length 5.
}
\usage{
gene_expression_tissues
}
\description{
A data set containing the tissues in the gene expressions data.
}
\keyword{datasets}
