% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_expressions}
\alias{gene_expressions}
\title{Gene Expression data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 755320 rows and 7 columns.
}
\usage{
gene_expressions
}
\description{
A data set containing the Log 2 Fold Change and P-Value for genes, across different strains, tissues, sexes, and ages.
}
\keyword{datasets}
