% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_expressions_labels}
\alias{gene_expressions_labels}
\title{Gene Expression labels}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7284 rows and 10 columns.
}
\usage{
gene_expressions_labels
}
\description{
A data set containing the labels for gene expressions. Only genes that are significant (up or downregulated) and with names that are less than 18 characters long are eligible for being labeled.
}
\keyword{datasets}
