% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_expressions_mouse_line_tissues}
\alias{gene_expressions_mouse_line_tissues}
\title{Gene expressions mouse line tissues}
\format{
An object of class \code{list} of length 3.
}
\usage{
gene_expressions_mouse_line_tissues
}
\description{
A data set containing the tissues available for each mouse line.
}
\keyword{datasets}
