% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_expressions_tissue}
\alias{gene_expressions_tissue}
\title{Gene Expression tissue}
\format{
An object of class \code{list} of length 6.
}
\usage{
gene_expressions_tissue
}
\description{
A data set containing the tissues available for each strain in the gene expression data.
}
\keyword{datasets}
