% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magora_corrplot.R
\name{magora_corrplot}
\alias{magora_corrplot}
\title{Correlation plot for nanostring versus AMPAD data}
\usage{
magora_corrplot(data, use_theme_sage = TRUE)
}
\arguments{
\item{data}{Input data (\code{\link{nanostring_for_plot}}}

\item{use_theme_sage}{Whether to use \code{\link[sagethemes]{theme_sage}}. Defaults to TRUE.}
}
\value{
A ggplot2 object.
}
\description{
Creates a correlation plot comparing nanostring data to human genetic data, faceted by age and cluster.
}
