% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magora_heatmap.R
\name{magora_heatmap}
\alias{magora_heatmap}
\title{Heatmap for gene expression data}
\usage{
magora_heatmap(data, log_foldchange_cutoff = 2.5)
}
\arguments{
\item{data}{Input data (\code{\link{gene_expressions}}, filtered by a tissue and one or more genes.}

\item{log_foldchange_cutoff}{Cutoff for log fold change, used in plot legend}
}
\value{
A ggplot2 object.
}
\description{
Creates a heatmap showing the log fold change and p-value of genes across sex, age, tissues and models.
}
