% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magora_volcano_plot.R
\name{magora_volcano_plot}
\alias{magora_volcano_plot}
\title{Volcano plot for gene expression data}
\usage{
magora_volcano_plot(
  data,
  data_labels,
  type = "ggplot2",
  facet = TRUE,
  save_name,
  sample_frac = 1
)
}
\arguments{
\item{data}{Gene expression data from \code{\link{gene_expressions}}, optionally filtered.}

\item{data_labels}{Labels for expression data from \code{\link{gene_expressions}}, optionally filtered. Only required if \code{type} is "ggplot2".}

\item{type}{Type of plot - one of "ggplot2" or "plotly". Defaults to "ggplot2".}

\item{facet}{Whether to facet the data by \code{sex} and \code{age}. Defaults to TRUE.}

\item{save_name}{A name that will be used for saving the plot. Only required / used when \code{type} is "plotly".}

\item{sample_frac}{The fraction of genes that are "not significant" that will be shown in the plot. Useful when there is a lot of data that slows down rendering. Defaults to 1.}
}
\description{
Volcano plot for gene expression data
}
