% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nanostring_for_plot}
\alias{nanostring_for_plot}
\title{Nanostring plotting data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 420 rows and 10 columns.
}
\usage{
nanostring_for_plot
}
\description{
A version of the data in \code{\link{nanostring}} specifically designed for plotting via \code{\link{magora_corrplot}}
}
\keyword{datasets}
