% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathology}
\alias{pathology}
\title{Pathology data}
\format{
An object of class \code{list} of length 3.
}
\usage{
pathology
}
\description{
A data set containing mouse phenotypes along with tissue, sex, age, and mouse line.
}
\keyword{datasets}
