% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathology_mouse_models}
\alias{pathology_mouse_models}
\title{Pathology mouse models}
\format{
An object of class \code{list} of length 2.
}
\usage{
pathology_mouse_models
}
\description{
A data set containing the variant and control for each mouse model in the pathology data.
}
\keyword{datasets}
