% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathology_tissue}
\alias{pathology_tissue}
\title{Pathology tissues}
\format{
An object of class \code{list} of length 16.
}
\usage{
pathology_tissue
}
\description{
A data set containing the tissues available for each mouse phenotype.
}
\keyword{datasets}
