% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phenotype_tissue}
\alias{phenotype_tissue}
\title{Phenotypes tissues}
\format{
An object of class \code{list} of length 15.
}
\usage{
phenotype_tissue
}
\description{
A data set containing the tissues available for each mouse phenotype.
}
\keyword{datasets}
