% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phenotypes}
\alias{phenotypes}
\title{Phenotypes data}
\format{
<<<<<<< HEAD
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2090 rows and 8 columns.
=======
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2096 rows and 10 columns.
>>>>>>> Clean up pathology names, add units
}
\usage{
phenotypes
}
\description{
A data set containing mouse phenotypes along with tissue, sex, age, and mouse line.
}
\keyword{datasets}
