% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_expressions.R
\name{plot_gene_expressions}
\alias{plot_gene_expressions}
\title{Plot gene expression data}
\usage{
plot_gene_expressions(data)
}
\arguments{
\item{data}{Input data (\code{\link{gene_expressions}} filtered by gene, mouse line, and tissue).}
}
\value{
A ggplot2 object.
}
\description{
Plot gene expression data
}
