% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phenotypes.R
\name{plot_phenotypes}
\alias{plot_phenotypes}
\title{Plot phenotype data}
\usage{
plot_phenotypes(data)
}
\arguments{
\item{data}{Input data ((\code{\link{phenotypes}} filtered by phenotype, mouse line, and tissue, with mouse lines expanded via \code{\link{expand_mouse_line_factor}}).}
}
\value{
A ggplot2 object.
}
\description{
Plot phenotype data
}
