% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactor_mouse_line.R
\name{refactor_mouse_line}
\alias{refactor_mouse_line}
\title{Refactor mouse line}
\usage{
refactor_mouse_line(data, mouse_line, phenotype_data = phenotypes)
}
\arguments{
\item{data}{Input data (the result of \code{\link{filter_pathology}} with \code{\link{phenotypes}} as input).}

\item{mouse_line}{Mouse line}

\item{phenotype_data}{Phenotype data to pull original mouse line factors from. Defaults to \code{link{phenotypes}}.}
}
\description{
Expand mouse line factor to include all experiment and control lines for the specified \code{mouse_line}, even if there is not data for both.
}
