% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regenerate_pvalue.R
\name{regenerate_pvalue}
\alias{regenerate_pvalue}
\title{Regenerate p-value from log10(pvalue)}
\usage{
regenerate_pvalue(data)
}
\arguments{
\item{data}{Gene expression data (\link{gene_expressions})}
}
\description{
Regenerated p-values from their log10 values, which are much easier to store.
}
