% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_mouse_line_factor.R
\name{expand_mouse_line_factor}
\alias{expand_mouse_line_factor}
\title{Expand mouse line factor}
\usage{
expand_mouse_line_factor(
  data,
  mouse_line_group,
  reference_data = magora::phenotypes
)
}
\arguments{
\item{data}{Input data (e.g. \code{\link{phenotypes}} filtered by phenotype, mouse line group, and tissue).}

\item{mouse_line_group}{Mouse line group(s) to expand factor for.}

\item{reference_data}{Reference data to pull original mouse line factor levels from. Defaults to \code{\link{phenotypes}}.}
}
\description{
Expand \code{mouse_line} factor levels to include all experiment and control lines for the specified \code{mouse_line_group}, even if there is not data for both. Used before \code{\link{magora_boxplot}} to ensure that facets are shown for all relevant mouse lines.
}
