% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pathology.R
\name{filter_pathology}
\alias{filter_pathology}
\title{Filter pathology data}
\usage{
filter_pathology(.data, .phenotype, .mouse_line, .tissue, na.rm = TRUE)
}
\arguments{
\item{.data}{Input data (e.g. \code{\link{phenotypes}})}

\item{.phenotype}{Phenotype to filter by. Keeps data where \code{.data[["Phenotype"]] == .phenotype}}

\item{.mouse_line}{Mouse lines to filter by. Keeps data where \code{.data[["mouse_line"]] \%in\% .mouse_line}}

\item{.tissue}{Tissue to filter by. Keeps data where \code{.data[["Tissue"]] == .tissue}}

\item{na.rm}{Whether to remove NA values of \code{data[["value"]]}. Defaults to TRUE.}
}
\description{
Filter pathology data to only include data from the specified \code{.phenotype}, \code{.mouse_line}, and \code{.tissue} values.
}
\examples{
phenotypes \%>\%
  filter_pathology(.phenotype = "Plaque #", .mouse_line = "BL6", .tissue = "hippocampus")
}
