% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_expressions}
\alias{gene_expressions}
\title{Gene expression data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1440 rows and 7 columns.
}
\usage{
gene_expressions
}
\description{
A data set containing mouse gene expressions along with sex, age, and mouse line.
}
\keyword{datasets}
