% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magora_boxplot.R
\name{magora_boxplot}
\alias{magora_boxplot}
\title{Boxplot for phenotype data}
\usage{
magora_boxplot(data, mouse_model_groups, use_theme_sage = TRUE)
}
\arguments{
\item{data}{Input data (\code{\link{pathology}} filtered by phenotype, mouse model, and tissue, with mouse models expanded via \code{\link{expand_mouse_model_factor_from_selection}}).}

\item{mouse_model_groups}{Mouse models}

\item{use_theme_sage}{Whether to use \code{\link[sagethemes]{theme_sage}}. Defaults to TRUE.}
}
\value{
A ggplot2 object.
}
\description{
Boxplot for phenotype data
}
