% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathology}
\alias{pathology}
\title{Pathology data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1964 rows and 12 columns.
}
\usage{
pathology
}
\description{
A data set containing mouse phenotypes along with tissue, sex, age, and mouse line.
}
\keyword{datasets}
