% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phenotypes.R
\name{plot_phenotypes}
\alias{plot_phenotypes}
\title{Plot phenotype data}
\usage{
plot_phenotypes(.data, .phenotype)
}
\arguments{
\item{.data}{Input data}

\item{.phenotype}{}
}
\value{

}
\description{
Plot phenotype data
}
