% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phenotypes.R
\name{plot_phenotypes}
\alias{plot_phenotypes}
\title{Plot phenotype data}
\usage{
plot_phenotypes(data)
}
\arguments{
\item{data}{Input data (the result of \code{\link{filter_pathology}} and \code{\link{expand_mouse_line_factor}} with \code{\link{phenotypes}} as input).}
}
\value{
A ggplot2 object.
}
\description{
Plot phenotype data
}
